feat(ranking): refactor RankingTabs and update styles
- Refactored RankingTabs component to a new file structure for better organization.
- Updated imports in ranking components to use new paths.
- Enhanced styles for ranking components by utilizing Colors constants for consistency.
- Removed the old RankingTabs component to streamline the codebase.
